% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{cellassign}
\alias{cellassign}
\title{Assign cells to known cell types}
\usage{
cellassign(exprs_obj, rho, s = NULL, X = NULL, data_type = c("RNAseq",
  "MS"), max_em_iter = 100, rel_tol = 0.001, multithread = FALSE,
  verbose = FALSE, bp_param = BiocParallel::bpparam(),
  sce_assay = "counts")
}
\description{
Assign cells to known cell types
}
