% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign_em.R
\name{cellassign_em}
\alias{cellassign_em}
\title{Assign cells to known cell types}
\usage{
cellassign_em(exprs_obj, rho, s = NULL, X = NULL, exprs_obj_known = NULL,
  s_known = NULL, X_known = NULL, known_types = NULL,
  use_priors = FALSE, prior_type = "regular", delta_log_prior_mean = 0,
  phi_log_prior_mean = 0, delta_log_prior_scale = 1,
  phi_log_prior_scale = 1, random_effects = FALSE, data_type = c("RNAseq",
  "MS"), n_batches = 1, rel_tol_adam = 1e-04, rel_tol_em = 1e-04,
  max_iter_adam = 1e+05, max_iter_em = 20, learning_rate = 0.1,
  verbose = TRUE, sce_assay = "counts", phi_type = "global",
  gamma_init = NULL, num_runs = 1)
}
\arguments{
\item{exprs_obj}{SingleCellExperiment object}

\item{rho}{Marker gene matrix (binary)}

\item{s}{Numeric vector of size factors}

\item{X}{Numeric matrix of auxiliary variables (NULL for none)}

\item{exprs_obj_known}{SingleCellExperiment object for labeled data (semisupervised)}

\item{s_known}{Size factors for labeled cells}

\item{X_known}{Auxiliary variables for labeled cells}
}
\description{
Assign cells to known cell types
}
