% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\docType{data}
\name{cellassign_env}
\alias{cellassign_env}
\title{Basilisk environment}
\format{An object of class \code{BasiliskEnvironment} of length 1.}
\usage{
cellassign_env
}
\description{
Basilisk environment
}
\keyword{local}
