% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{cellprobs}
\alias{cellprobs}
\alias{cellprobs.cellassign}
\title{Get the cell assignment probabilities of a \code{cellassign} fit}
\usage{
cellprobs(x)

\method{cellprobs}{cellassign}(x)
}
\arguments{
\item{x}{An object of class \code{cellassign} returned by a call to \code{cellassign(...)}}
}
\value{
A cell by celltype matrix with assignment probabilities
}
\description{
Get the MLE cell type assignment probabilities for each cell
}
\examples{
data(example_cellassign_fit)
cellprobs(example_cellassign_fit)
}
