% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference.R
\name{clone_assignment}
\alias{clone_assignment}
\title{Computes map clone assignment given EM object}
\usage{
clone_assignment(em)
}
\arguments{
\item{em}{List returned by \code{inference_em}}
}
\value{
A vector of maximum likelihood clone assignments
}
\description{
Computes map clone assignment given EM object
}
\keyword{internal}

