% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-tensorflow.R
\name{construct_tf_formulas}
\alias{construct_tf_formulas}
\title{construct cellassign formulas for optimization (Raw expression matrix. Negative binomial distribution hypothesis)}
\usage{
construct_tf_formulas(G, C, P, B, minY, maxY, shrinkage, min_delta,
  dirichlet_concentration, random_seed, tf, tfd)
}
\value{
A list of variables for EM optimization
}
\description{
construct cellassign formulas for optimization (Raw expression matrix. Negative binomial distribution hypothesis)
}
\keyword{internal}
