% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-tensorflow.R
\name{estimate_base_expression}
\alias{estimate_base_expression}
\title{Estimate base expression for each marker gene}
\usage{
estimate_base_expression(Y)
}
\arguments{
\item{Y}{normalized expression matrix with all genes}
}
\value{
a numeric list of base expression estimates
}
\description{
Given the normalized expression matrix, estimate base expression for each gene
by fitting normalized expression data to gaussian mixture models
}
\details{
This function takes the normalized expression matrix and return a list of base expression estimates
The input list should be the complete normalized expression matrix
}
