% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\docType{data}
\name{example_cellassign_fit}
\alias{example_cellassign_fit}
\title{Example cellassign fit}
\format{An object of class \code{cellassign} of length 3.}
\usage{
example_cellassign_fit
}
\description{
An example fit of calling \code{cellassign} on both
\code{example_marker_mat} and \code{example_sce}
}
\examples{
data(example_cellassign_fit)
}
\seealso{
example_cellassign_fit
}
\keyword{datasets}
