% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\docType{data}
\name{example_rho}
\alias{example_rho}
\title{Example cell marker matrix}
\format{An object of class \code{matrix} with 10 rows and 2 columns.}
\usage{
example_rho
}
\description{
An example matrix for 10 genes and 2 cell types showing the membership
of marker genes to cell types
}
\examples{
data(example_rho)
}
\seealso{
example_cellassign_fit
}
\keyword{datasets}
