% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example SingleCellExperiment}
\format{An object of class \code{SingleCellExperiment} with 10 rows and 500 columns.}
\usage{
example_sce
}
\description{
An example \code{SingleCellExperiment} for 10 marker genes and 500 cells.
}
\examples{
data(example_sce)
}
\seealso{
example_cellassign
}
\keyword{datasets}
