% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\docType{data}
\name{example_TME_markers}
\alias{example_TME_markers}
\title{Example tumour microevironment markers}
\format{An object of class \code{list} of length 2.}
\usage{
example_TME_markers
}
\description{
A set of example marker genes for commonly profiling the
human tumour mircoenvironment
}
\examples{
data(example_TME_markers)
}
\keyword{datasets}
