% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_expression_matrix}
\alias{extract_expression_matrix}
\title{Extract expression matrix from expression object}
\usage{
extract_expression_matrix(exprs_obj, sce_assay = "counts")
}
\description{
Extract expression matrix from expression object
}
\keyword{internal}
