% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_delta.R
\name{fit_delta}
\alias{fit_delta}
\title{Fit delta based on known annotations}
\usage{
fit_delta(exprs_obj, rho, s = NULL, X = NULL, known_types = NULL,
  data_type = c("RNAseq", "MS"), n_batches = 1, rel_tol_adam = 1e-04,
  rel_tol_em = 1e-04, max_iter_adam = 1e+05, max_iter_em = 20,
  learning_rate = 0.1, verbose = TRUE, sce_assay = "counts")
}
\description{
Fit delta based on known annotations
}
