% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mle_cell_type}
\alias{get_mle_cell_type}
\title{Get MLE estimates of type of each cell}
\usage{
get_mle_cell_type(gamma)
}
\value{
A vector of MLE cell types, where the names are
taken from the column names of the input matrix
}
\description{
Get MLE estimates of type of each cell
}
\keyword{internal}
