% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\docType{data}
\name{holik_data}
\alias{holik_data}
\title{Example bulk RNA-seq data}
\format{An object of class \code{list} of length 2.}
\usage{
holik_data
}
\description{
An example bulk RNA-seq dataset from Holik et al. Nucleic Acids Research 2017 to
demonstrate deriving marker genes
}
\examples{
data(holik_data)
}
\keyword{datasets}
