% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_X}
\alias{initialize_X}
\title{Create X matrix}
\usage{
initialize_X(X, N, verbose = FALSE)
}
\value{
A cleaned covariate matrix given the input provided by the user
}
\description{
Create X matrix
}
\keyword{internal}
