% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-tensorflow.R
\name{mle_tensorflow}
\alias{mle_tensorflow}
\title{cellassign inference in tensorflow, semi-supervised version}
\usage{
mle_tensorflow(Y, rho, s, X, G, C, N, P, gamma, verbose = FALSE,
  rel_tol_adam = 1e-04, rel_tol_em = 1e-04, max_iter_adam = 1e+05,
  max_iter_em = 20, learning_rate = 0.1)
}
\description{
cellassign inference in tensorflow, semi-supervised version
}
