% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{mleparams}
\alias{mleparams}
\alias{mleparams.cellassign}
\title{Get the MLE parameter list of a \code{cellassign} fit}
\usage{
mleparams(x)

\method{mleparams}{cellassign}(x)
}
\arguments{
\item{x}{An object of class \code{cellassign} returned
by a call to \code{cellassign(...)}}
}
\value{
A list of MLE parameter estimates from cellassign
}
\description{
Get the MLE parameter list of a \code{cellassign} fit
}
\examples{
data(example_cellassign_fit)
mleparams(example_cellassign_fit)
}
