% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{offset_rho}
\alias{offset_rho}
\title{Offset rho matrix with some uncertainty}
\usage{
offset_rho(x, offset = 0.05)
}
\value{
A rho matrix offset inwards by some value
}
\description{
Offset rho matrix with some uncertainty
}
\keyword{internal}
