% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference.R
\name{p_pi}
\alias{p_pi}
\title{Computes gamma_{nc} = p(pi_n = c), returning
N by C matrix of responsibilities}
\usage{
p_pi(data, params)
}
\arguments{
\item{data}{Input data}

\item{params}{Model parameters}
}
\value{
The probability that each cell belongs to each cell type, as a matrix
}
\description{
Computes gamma_{nc} = p(pi_n = c), returning
N by C matrix of responsibilities
}
\keyword{internal}
