% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference-nb.R
\name{p_pi_nb}
\alias{p_pi_nb}
\title{Computes gamma_{nc} = p(pi_n = c), returning
N by C matrix of responsibilities}
\usage{
p_pi_nb(data, params)
}
\arguments{
\item{data}{Input data}

\item{params}{Model parameters}
}
\value{
The probability that each cell belongs to each cell type, as a matrix

TODO: make this p_pi for each cell in order to parallelize
}
\description{
Computes gamma_{nc} = p(pi_n = c), returning
N by C matrix of responsibilities
}
\keyword{internal}
