% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference-nb.R
\name{p_pi_nb2}
\alias{p_pi_nb2}
\title{This is the same as p_pi_nb but parallel over cells,
which should speed up computation as datasets become large}
\usage{
p_pi_nb2(data, params, multithread, bp_param)
}
\description{
This is the same as p_pi_nb but parallel over cells,
which should speed up computation as datasets become large
}
