% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{print.cellassign}
\alias{print.cellassign}
\title{Print a \code{cellassign} fit}
\usage{
\method{print}{cellassign}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cellassign}}

\item{...}{Additional arguments (unused)}
}
\value{
Prints a structured representation of the \code{cellassign}
}
\description{
Print a \code{cellassign} fit
}
\examples{
data(example_cellassign)
print(example_cellassign)

}
