% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{print.cellassign_fit}
\alias{print.cellassign_fit}
\title{Print a \code{cellassign} fit}
\usage{
\method{print}{cellassign_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cellassign_fit}}

\item{...}{Additional arguments (unused)}
}
\value{
Prints a structured representation of the \code{cellassign_fit}
}
\description{
Print a \code{cellassign} fit
}
\examples{
data(example_cellassign_fit)
print(example_cellassign_fit)

}
