% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference-nb.R
\name{Q_g_nb}
\alias{Q_g_nb}
\title{Q(theta | theta^t)}
\usage{
Q_g_nb(pars, y, rho, gamma, data)
}
\arguments{
\item{pars}{Parameters to optimize}

\item{y}{Gene expression}

\item{gamma}{Responsibility terms (expectation of clone assignments), N-by-C}

\item{data}{Data}
}
\description{
Q(theta | theta^t)
}
\keyword{internal}
