% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cellassign}
\alias{simulate_cellassign}
\title{Simulate from the cellassign model}
\usage{
simulate_cellassign(rho, s, pi, delta, phi, beta, X = NULL)
}
\arguments{
\item{rho}{A gene by cell type binary matrix relating markers to cell types}

\item{s}{A vector of cell-specific size factors}

\item{pi}{An ordinal vector relating each cell to its true marker type assignment}

\item{delta}{Gene by cell type matrix delta (all entries with corresponding zeros
in rho will be ignored)}

\item{phi}{Gene by cell matrix of dispersion parameters}

\item{beta}{A gene by covariate vector of coefficients - the first column
should correspond to the intercept (baseline expression) values}

\item{X}{A cell by covariate matrix of covariates - the intercept column will
always be added.}
}
\value{
An N by G matrix of simulated counts
}
\description{
Simulate RNA-seq counts from the cell-assign model
}
\details{
The number of genes, cells, and cell types is automatically
inferred from the dimensions of rho (gene by cell-type) and
s (vector of length number of cells). The specification of X
is optional - a column of ones will always be added as an intercept.
}
