% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference-nb.R
\name{slice_parameters_nb}
\alias{slice_parameters_nb}
\title{Extracts parameters from array}
\usage{
slice_parameters_nb(pars, rho, X)
}
\description{
Extracts parameters from array
}
\keyword{internal}
