% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_genes}
\alias{validate_genes}
\title{Extract expression matrix from expression object}
\usage{
validate_genes(Y, rho)
}
\value{
The expression matrix and marker gene matrix, with only expressed genes, for input to \code{cellassign}
}
\description{
Extract expression matrix from expression object
}
\keyword{internal}
