% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb-tensorflow.R
\name{vb_tensorflow}
\alias{vb_tensorflow}
\title{cellassign inference in tensorflow, VB version}
\usage{
vb_tensorflow(Y, rho, s, X, G, C, N, P, B = 10, S = 10,
  rn_inputs = NULL, rn_layers = c(32, 32), verbose = FALSE,
  n_batches = 1, tol = 1e-04, max_epochs = 5000,
  period_length = 100, num_tol_periods = 2, learning_rate = 0.01,
  random_seed = NULL, min_delta = 2, dirichlet_concentration = rep(1,
  C))
}
\value{
A list of cell type calls, parameter estimates,
and ELBOs during optimization.
}
\description{
cellassign inference in tensorflow, VB version
}
\keyword{internal}
