% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{cellassign}
\alias{cellassign}
\title{Assign cells to known cell types}
\usage{
cellassign(exprs_obj, rho, s = NULL, X = NULL, exprs_obj_known = NULL,
  s_known = NULL, X_known = NULL, delta_alpha_prior = NULL,
  delta_beta_prior = NULL, known_types = NULL, data_type = c("RNAseq",
  "MS"), n_batches = 1, rel_tol_adam = 1e-04, rel_tol_em = 1e-04,
  max_iter_adam = 1e+05, max_iter_em = 20, max_adam_epochs = 5000,
  period_epochs = 400, no_change_rel_thres = 1e-05, no_change_periods = 2,
  num_vb_samples = 10, learning_rate = 0.1, verbose = TRUE,
  sce_assay = "counts", phi_type = "global", gamma_init = NULL,
  num_runs = 1, num_hidden_nodes_vb = 50, inference_method = "EM")
}
\arguments{
\item{exprs_obj}{SingleCellExperiment object}

\item{rho}{Marker gene matrix (binary)}

\item{s}{Numeric vector of size factors}

\item{X}{Numeric matrix of auxiliary variables (NULL for none)}

\item{exprs_obj_known}{SingleCellExperiment object for labeled data (semisupervised)}

\item{s_known}{Size factors for labeled cells}

\item{X_known}{Auxiliary variables for labeled cells}
}
\description{
Assign cells to known cell types
}
