% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{cellassign}
\alias{cellassign}
\title{Annotate cells to cell types using cellassign}
\usage{
cellassign(exprs_obj, rho, s = NULL, X = NULL, control_pct = NULL,
  exprs_obj_known = NULL, s_known = NULL, X_known = NULL,
  control_pct_known = NULL, known_types = NULL, B = 10,
  pct_mito = NULL, mito_rho = NULL, use_priors = FALSE,
  use_mito = FALSE, prior_type = "regular",
  delta_log_prior_mean = NULL, delta_log_prior_scale = 1,
  delta_variance_prior = FALSE, random_effects = FALSE,
  data_type = c("RNAseq", "MS"), n_batches = 1, rel_tol_adam = 1e-04,
  rel_tol_em = 1e-04, max_iter_adam = 1e+05, max_iter_em = 20,
  learning_rate = 0.1, verbose = TRUE, sce_assay = "counts",
  gamma_init = NULL, num_runs = 1, em_convergence_thres = 1e-05,
  min_delta = log(2))
}
\arguments{
\item{exprs_obj}{Either a matrix representing gene expression counts or a \code{SummarizedExperiment}.
See details.}

\item{rho}{TODO}

\item{s}{Numeric vector of cell size factors}

\item{X}{Numeric matrix of external covariates. See details.}

\item{exprs_obj_known}{SingleCellExperiment object for labeled data (semisupervised)}

\item{s_known}{Size factors for labeled cells}

\item{X_known}{Auxiliary variables for labeled cells}
}
\description{
Automatically annotate cells to known types based on the expression patterns of
a priori known marker genes.
}
\details{
\strong{Input format}
\code{exprs_obj} should be either a \code{SummarizedExperiment} (we recommend the
\code{SingleCellExperiment} package) or a cell (row) by gene (column) matrix of
\emph{raw} RNA-seq counts (do \strong{not} log-transform or otherwise normalize).

\strong{Cell size factors}
If the cell size factors \code{s} are not provided they are computed using the
\code{computeSumFactors} function from the \code{scran} package.

\strong{Covariates}
If \code{X} is not \code{NULL} then it should be an \code{N} by \code{P} matrix
of covariates for \code{N} cells and \code{P} covariates. Such a matrix would typically
be returned by a call to \code{model.matrix} \strong{with no intercept}. It is also highly
recommended that any numerical (ie non-factor or one-hot-encoded) covariates be standardized
to have mean 0 and standard deviation 1.
}
