% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{cellassign}
\alias{cellassign}
\title{Assign cells to known cell types}
\usage{
cellassign(exprs_obj, rho, s = NULL, X = NULL, data_type = c("RNAseq",
  "MS"), n_batches = 1, rel_tol_adam = 1e-04, rel_tol_em = 1e-04,
  max_iter_adam = 1e+05, max_iter_em = 20, learning_rate = 0.1,
  verbose = TRUE, sce_assay = "counts")
}
\description{
Assign cells to known cell types
}
