% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign-inference.R
\name{cellassign_inference}
\alias{cellassign_inference}
\title{Expectation-maximization for cellassign}
\usage{
cellassign_inference(Y, rho, X = NULL, max_em_iter = 100, rel_tol = 0.001,
  multithread = FALSE, verbose = FALSE)
}
\arguments{
\item{Y}{Input matrix of expression counts for N cells (rows)
by G genes (columns)}

\item{rho}{Binary matrix of G rows and C columns where
rho_{gc} = 1 if gene g is a marker for cell type c}

\item{X}{An N by (P-1) matrix of covariates *without* an intercept}

\item{max_em_iter}{Maximum number of EM iterations to perform}

\item{rel_tol}{Tolerance below which EM algorithm is considered converged}
}
\description{
Expectation-maximization for cellassign
}

