% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign_vb.R
\name{cellassign_vb}
\alias{cellassign_vb}
\title{Assign cells to known cell types}
\usage{
cellassign_vb(exprs_obj, rho, s = NULL, X = NULL,
  delta_alpha_prior = NULL, delta_beta_prior = NULL, known_types = NULL,
  data_type = c("RNAseq", "MS"), n_batches = 1, max_adam_epochs = 5000,
  period_epochs = 400, no_change_rel_thres = 1e-05, no_change_periods = 2,
  num_vb_samples = 10, learning_rate = 0.1, verbose = TRUE,
  sce_assay = "counts", phi_type = "global", num_hidden_nodes_vb = 50,
  variance_multiplier = 10, marker_multiplier = 4)
}
\arguments{
\item{exprs_obj}{SingleCellExperiment object}

\item{rho}{Marker gene matrix (binary)}

\item{s}{Numeric vector of size factors}

\item{X}{Numeric matrix of auxiliary variables (NULL for none)}

\item{exprs_obj_known}{SingleCellExperiment object for labeled data (semisupervised)}

\item{s_known}{Size factors for labeled cells}

\item{X_known}{Auxiliary variables for labeled cells}
}
\description{
Assign cells to known cell types
}
