% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{celltypes}
\alias{celltypes}
\alias{celltypes.cellassign}
\title{Get the cell type assignments of a \code{cellassign} fit}
\usage{
celltypes(x)

\method{celltypes}{cellassign}(x)
}
\arguments{
\item{x}{An object of class \code{cellassign} returned by a call to \code{cellassign(...)}}
}
\value{
A character vector with the MLE cell type for each cell
}
\description{
Get the MLE cell type estimates for each cell
}
\examples{
data(example_cellassign_fit)
celltypes(example_cellassign_fit)
}
