% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-tensorflow.R
\name{inference_tensorflow}
\alias{inference_tensorflow}
\title{cellassign inference in tensorflow, semi-supervised version}
\usage{
inference_tensorflow(Y, rho, s, X, G, C, N, P, control_pct,
  pct_mito = rep(0, nrow(Y)), mito_rho = rep(0, C), Y0, s0, X0, N0, P0,
  gamma0, control_pct0, B = 10, use_priors, use_mito,
  prior_type = "regular", delta_log_prior_mean = NULL,
  delta_log_prior_scale = 1, delta_variance_prior = FALSE,
  random_effects = FALSE, verbose = FALSE, n_batches = 1,
  rel_tol_adam = 1e-04, rel_tol_em = 1e-04, max_iter_adam = 1e+05,
  max_iter_em = 20, learning_rate = 1e-04, gamma_init = NULL,
  random_seed = NULL, em_convergence_thres = 1e-05,
  min_delta = log(2))
}
\description{
cellassign inference in tensorflow, semi-supervised version
}
