% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb-tensorflow.R
\name{vb_tensorflow}
\alias{vb_tensorflow}
\title{Variational Bayes version of cellassign}
\usage{
vb_tensorflow(rho_dat, Y_dat, Y_std_dat, s_dat, X_dat, delta_alpha_dat,
  delta_beta_dat, G, C, N, P, S = 10, verbose = FALSE,
  num_hidden_nodes = 50, learning_rate = 0.001, n_batches = 1,
  max_adam_epoch = 5000, period_epochs = 400, no_change_rel_thres = 1e-05,
  no_change_periods = 2)
}
\arguments{
\item{rho_dat}{Marker gene matrix (binary)}

\item{Y_dat}{Expression matrix}

\item{Y_std_dat}{Expression matrix, standardized}

\item{X_dat}{Auxiliary variable matrix}

\item{delta_alpha_dat}{Prior over alpha parameter for delta}

\item{delta_beta_dat}{Prior over beta parameter for delta}

\item{G}{Number of genes}

\item{C}{Number of cell types}

\item{N}{Number of cells}

\item{P}{Number of auxiliary variables}

\item{S}{Number of VB samples to take}

\item{num_hidden_nodes}{Number of hidden nodes in neural network}

\item{learning_rate}{Learning rate}

\item{n_batches}{Number of batches to process data in}

\item{max_adam_epoch}{Maximum number of ADAM epochs to run}

\item{period_epochs}{Message period (if on verbose)}

\item{no_change_rel_thres}{Threshold below which relative change in ELBO is considered to be 0}

\item{no_change_periods}{Number of periods with no change before stopping}
}
\description{
Variational Bayes version of cellassign
}
